requirements.txt
Flask
faiss-cpu
numpy
matplotlib
sentence-transformers
scikit-learn
groq